refactor(semaphore): remove IDatabaseSemaphoreAdapter contracts and fix import paths
