+++ src/event-bus/implementations/derivables/event-bus/is-event-bus.test.ts
@@ -10,7 +10,7 @@ import {
     type Unsubscribe,
 } from "@/event-bus/contracts/_module.js";
 import { isEventBus } from "@/event-bus/implementations/derivables/event-bus/is-event-bus.js";
-import { type IReadableContext } from "@/execution-context/contracts/execution-context.contract.js";
+import { type IReadableContext } from "@/execution-context/contracts/_module.js";
 import { type OneOrArray } from "@/utilities/_module.js";
 
 describe("function: isEventBus", () => {
